--- src/main/java/com/automated/trading/stock/StockManager/community/contact/ContactResponseType.java
@@ -1,5 +1,0 @@
-package com.automated.trading.stock.StockManager.community.contact;
-
-public enum ContactResponseType {
-    accept, reject, hold;
-}
