+++ src/main/java/com/automated/trading/stock/StockManager/community/contact/domain/ContactResponseType.java
@@ -1,0 +1,5 @@
+package com.automated.trading.stock.StockManager.community.contact.domain;
+
+public enum ContactResponseType {
+    notChecked, accepted, rejected, hold;
+}
